requirements.txt
asgiref==3.5.0
async-generator==1.10
attrs==21.4.0
autopep8==1.6.0
beautifulsoup4==4.10.0
boto3==1.21.4
botocore==1.24.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.1
cycler==0.11.0
defusedxml==0.7.1
demjson==2.2.4
Deprecated==1.2.13
diff-match-patch==20200713
dj-database-url==0.5.0
Django==4.0.2
django-clearcache==1.2.1
django-crispy-forms==1.14.0
django-debug-toolbar==3.2.4
django-environ==0.8.1
django-grappelli==3.0.3
django-heroku==0.3.1
django-import-export==2.7.1
django-storages==1.12.3
et-xmlfile==1.1.0
fonttools==4.29.1
gunicorn==20.1.0
h11==0.13.0
idna==3.3
jmespath==0.10.0
kiwisolver==1.3.2
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.9
outcome==1.1.0
packaging==21.3
Pillow==9.0.1
psycopg2==2.9.3
pycodestyle==2.8.0
pycparser==2.21
pyOpenSSL==22.0.0
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
redis==4.1.4
requests==2.27.1
s3transfer==0.5.1
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
sqlparse==0.4.2
tablib==3.2.0
toml==0.10.2
trio==0.20.0
trio-websocket==0.9.2
tzdata==2021.5
urllib3==1.26.8
whitenoise==6.0.0
wrapt==1.14.0
wsproto==1.0.0
xlrd==2.0.1
xlwt==1.3.0
xmltodict==0.12.0
